Controllers and repositories

+++ src/main/java/com/gymmanagement/gymmanagement/membership_ticket/MembershipTicketRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MembershipTicketRepository extends JpaRepository<MembershipTicketEntity, Long> {
+public interface MembershipTicketRepository extends JpaRepository<MembershipTicketEntity, Integer> {
